Add support for MSVC

+++ src/libshibatchdsp/ssrcsoxr.cpp
@@ -316,7 +316,7 @@ struct ssrc_soxr *ssrc_soxr_create(double input_rate, double output_rate, unsign
     vector<shared_ptr<StageOutlet<float>>> out(num_channels);
 
     for(unsigned i=0;i<num_channels;i++) {
-      auto ssrc = make_shared<SSRC<float>>(xifier->getOutlet(i), input_rate, output_rate, q.log2dftfilterlen, q.aa, q.guard);
+      auto ssrc = make_shared<SSRC<float>>(xifier->getOutlet(i), (int64_t)input_rate, (int64_t)output_rate, q.log2dftfilterlen, q.aa, q.guard);
       thiz->delay = ssrc->getDelay();
       out[i] = ssrc;
     }
@@ -386,7 +386,7 @@ ssrc_soxr_error_t ssrc_soxr_clear(struct ssrc_soxr *thiz) {
     vector<shared_ptr<StageOutlet<float>>> out(thiz->num_channels);
 
     for(unsigned i=0;i<thiz->num_channels;i++) {
-      auto ssrc = make_shared<SSRC<float>>(xifier->getOutlet(i), thiz->input_rate, thiz->output_rate,
+      auto ssrc = make_shared<SSRC<float>>(xifier->getOutlet(i), (int64_t)thiz->input_rate, (int64_t)thiz->output_rate,
 					   thiz->qspec.log2dftfilterlen, thiz->qspec.aa, thiz->qspec.guard);
       thiz->delay = ssrc->getDelay();
       out[i] = ssrc;
